Update v0.6
- Add updateData() function -> handlers.go
- Add garbageTypes() function & recyclingMethods() function -> config.go
- Rename function.go -> handlers.go
- Rename globals.go -> config.go
- Rename utils.go -> helpers.go
- Cleanup handlers.go
